requirements.txt
Gooey
pillow
pyinstaller